feat: accept fractional experience (months or years) at intake
Carriers in our pool will hire from 3 months of Class A experience, but
intake forced a whole-number year. drivers.years_held and otr_years
become numeric(5,2) so 3 months = 0.25 stores cleanly. The intake form
now shows a number input with a months/years unit selector; the form
converts to years on submit, and the matching engine still computes
months via Math.round(yearsHeld × 12). Zod no longer requires an
integer.

Migration 0004 alters the columns in place.

+++ src/lib/intake-schema.ts
@@ -76,14 +76,16 @@ export const intakeSchema = z.object({
     .string()
     .trim()
     .regex(/^\d{5}$/, "Use a 5-digit US zip"),
-  yearsHeld: z.coerce.number().int().min(0).max(60),
+  // Stored as years (decimal) so 3 months = 0.25. The intake form lets the
+  // driver enter a value in months or years; submission converts to years.
+  yearsHeld: z.coerce.number().min(0).max(60),
 
   // Step 2: Experience + Equipment
   equipmentRun: z
     .array(z.string())
     .min(1, "Pick at least one equipment type you've actually driven"),
   endorsements: z.array(z.string()).default([]),
-  otrYears: z.coerce.number().int().min(0).max(60).default(0),
+  otrYears: z.coerce.number().min(0).max(60).default(0),
 
   // Step 3: Preferences
   desiredEquipment: z
